News Article/requirements.txt
pandas
matplotlib
seaborn
nltk
tqdm
scikit-learn
scipy
numpy
imbalanced-learn
Django
djangorestframework
django-extensions